First attempt at moving fucntion to vision
