+++ components/camera/transformpipeline/pipeline_test.go
@@ -50,10 +50,6 @@ func TestTransformPipelineColor(t *testing.T) {
 	_, err = color.NextPointCloud(context.Background())
 	test.That(t, err, test.ShouldNotBeNil)
 	test.That(t, err, test.ShouldWrap, transform.ErrNoIntrinsics)
-	_, err = color.Projector(context.Background())
-	test.That(t, err, test.ShouldWrap, transform.ErrNoIntrinsics)
-	test.That(t, color.Close(context.Background()), test.ShouldBeNil)
-	test.That(t, source.Close(context.Background()), test.ShouldBeNil)
 }
 
 func TestTransformPipelineDepth(t *testing.T) {
@@ -94,9 +90,9 @@ func TestTransformPipelineDepth(t *testing.T) {
 	test.That(t, err, test.ShouldBeNil)
 	test.That(t, outImg.Bounds().Dx(), test.ShouldEqual, 40)
 	test.That(t, outImg.Bounds().Dy(), test.ShouldEqual, 30)
-	prop, err := depth.Projector(context.Background())
+	prop, err := depth.Properties(context.Background())
 	test.That(t, err, test.ShouldBeNil)
-	test.That(t, prop, test.ShouldResemble, intrinsics)
+	test.That(t, prop.IntrinsicParams, test.ShouldResemble, intrinsics)
 	outPc, err := depth.NextPointCloud(context.Background())
 	test.That(t, err, test.ShouldNotBeNil)
 	test.That(t, err.Error(), test.ShouldContainSubstring, "not defined for last videosource")
@@ -209,10 +205,10 @@ func TestPipeIntoPipe(t *testing.T) {
 	test.That(t, err, test.ShouldBeNil)
 	test.That(t, outImg.Bounds().Dx(), test.ShouldEqual, 128)
 	test.That(t, outImg.Bounds().Dy(), test.ShouldEqual, 72)
-	prop, err := pipe1.Projector(context.Background())
+	prop, err := pipe1.Properties(context.Background())
 	test.That(t, err, test.ShouldBeNil)
-	test.That(t, prop.(*transform.PinholeCameraIntrinsics).Width, test.ShouldEqual, 128)
-	test.That(t, prop.(*transform.PinholeCameraIntrinsics).Height, test.ShouldEqual, 72)
+	test.That(t, prop.IntrinsicParams.Width, test.ShouldEqual, 128)
+	test.That(t, prop.IntrinsicParams.Height, test.ShouldEqual, 72)
 	// transform pipeline into pipeline
 	pipe2, err := newTransformPipeline(context.Background(), pipe1, transform2, r, logger)
 	test.That(t, err, test.ShouldBeNil)
@@ -221,10 +217,10 @@ func TestPipeIntoPipe(t *testing.T) {
 	test.That(t, outImg.Bounds().Dx(), test.ShouldEqual, 10)
 	test.That(t, outImg.Bounds().Dy(), test.ShouldEqual, 20)
 	test.That(t, err, test.ShouldBeNil)
-	prop, err = pipe2.Projector(context.Background())
+	prop, err = pipe2.Properties(context.Background())
 	test.That(t, err, test.ShouldBeNil)
-	test.That(t, prop.(*transform.PinholeCameraIntrinsics).Width, test.ShouldEqual, 10)
-	test.That(t, prop.(*transform.PinholeCameraIntrinsics).Height, test.ShouldEqual, 20)
+	test.That(t, prop.IntrinsicParams.Width, test.ShouldEqual, 10)
+	test.That(t, prop.IntrinsicParams.Height, test.ShouldEqual, 20)
 	// Close everything
 	test.That(t, pipe2.Close(context.Background()), test.ShouldBeNil)
 	test.That(t, pipe1.Close(context.Background()), test.ShouldBeNil)
